+++ packages/expo-camera/src/ExpoCamera.web.tsx
@@ -1,15 +1,19 @@
 import { CodedError } from 'expo-modules-core';
-import * as React from 'react';
+import React, {
+  type PropsWithChildren,
+  forwardRef,
+  useRef,
+  useMemo,
+  useImperativeHandle,
+  type ComponentProps,
+  type Ref,
+} from 'react';
 import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
 import createElement from 'react-native-web/dist/exports/createElement';
 
+import { CameraNativeProps, CameraCapturedPicture } from './Camera.types';
 import CameraManager from './ExpoCameraManager.web';
-import {
-  CameraCapturedPicture,
-  CameraNativeProps,
-  CameraPictureOptions,
-  CameraType,
-} from './legacy/Camera.types';
+import { CameraPictureOptions, CameraType } from './legacy/Camera.types';
 import { capture } from './web/WebCameraUtils';
 import { PictureSizes } from './web/WebConstants';
 import { useWebCameraStream } from './web/useWebCameraStream';
@@ -22,14 +26,14 @@ export interface ExponentCameraRef {
   pausePreview: () => Promise<void>;
 }
 
-const ExponentCamera = React.forwardRef(
+const ExponentCamera = forwardRef(
   (
-    { type, poster, ...props }: CameraNativeProps & { children?: React.ReactNode },
-    ref: React.Ref<ExponentCameraRef>
+    { facing, poster, ...props }: PropsWithChildren<CameraNativeProps>,
+    ref: Ref<ExponentCameraRef>
   ) => {
-    const video = React.useRef<HTMLVideoElement | null>(null);
+    const video = useRef<HTMLVideoElement | null>(null);
 
-    const native = useWebCameraStream(video, type as CameraType, props, {
+    const native = useWebCameraStream(video, facing as CameraType, props, {
       onCameraReady() {
         if (props.onCameraReady) {
           props.onCameraReady();
@@ -38,30 +42,27 @@ const ExponentCamera = React.forwardRef(
       onMountError: props.onMountError,
     });
 
-    const isQRScannerEnabled = React.useMemo<boolean>(() => {
-      return !!(
-        props.barCodeScannerSettings?.barCodeTypes?.includes('qr') && !!props.onBarCodeScanned
+    const isQRScannerEnabled = useMemo<boolean>(() => {
+      return Boolean(
+        props.barcodeScannerSettings?.barcodeTypes?.includes('qr') && !!props.onBarcodeScanned
       );
-    }, [props.barCodeScannerSettings?.barCodeTypes, props.onBarCodeScanned]);
+    }, [props.barcodeScannerSettings?.barcodeTypes, props.onBarcodeScanned]);
 
     useWebQRScanner(video, {
-      interval: props.barCodeScannerSettings?.interval,
+      interval: 300,
       isEnabled: isQRScannerEnabled,
       captureOptions: { scale: 1, isImageMirror: native.type === CameraType.front },
       onScanned(event) {
-        if (props.onBarCodeScanned) {
-          props.onBarCodeScanned(event);
+        if (props.onBarcodeScanned) {
+          props.onBarcodeScanned(event);
         }
       },
-      // onError: props.onMountError,
     });
 
-    // const [pause, setPaused]
-
-    React.useImperativeHandle(
+    useImperativeHandle(
       ref,
       () => ({
-        async getAvailablePictureSizes(ratio: string): Promise<string[]> {
+        async getAvailablePictureSizes(): Promise<string[]> {
           return PictureSizes;
         },
         async takePicture(options: CameraPictureOptions): Promise<CameraCapturedPicture> {
@@ -107,7 +108,7 @@ const ExponentCamera = React.forwardRef(
     // Because we don't support recording video in the browser we don't need the user to give microphone permissions.
     const isMuted = true;
 
-    const style = React.useMemo<StyleProp<ViewStyle>>(() => {
+    const style = useMemo<StyleProp<ViewStyle>>(() => {
       const isFrontFacingCamera = native.type === CameraManager.Type.front;
       return [
         StyleSheet.absoluteFill,
@@ -126,7 +127,6 @@ const ExponentCamera = React.forwardRef(
           playsInline
           muted={isMuted}
           poster={poster}
-          // webkitPlaysinline
           pointerEvents={props.pointerEvents}
           ref={video}
           style={style}
@@ -139,15 +139,15 @@ const ExponentCamera = React.forwardRef(
 
 export default ExponentCamera;
 
-const Video = React.forwardRef(
+const Video = forwardRef(
   (
-    props: React.ComponentProps<typeof View> & {
+    props: ComponentProps<typeof View> & {
       autoPlay?: boolean;
       playsInline?: boolean;
       muted?: boolean;
       poster?: string;
     },
-    ref: React.Ref<HTMLVideoElement>
+    ref: Ref<HTMLVideoElement>
   ) => createElement('video', { ...props, ref })
 );